requirements.txt
python-dotenv>=1.0.0
langchain-openai>=0.1.0
langgraph>=0.0.35
openai>=1.16.0

